// src/controls/imagePicker/models/ISearchImagesResult.ts
 

export interface ISearchImagesResult {
  driveId: string;
  title: string;
  path: string;
  filename: string;
  fileExtension: string;
  fileType: string;
  created: string;
  author: string;
  lastModifiedTime: string;
  editorOwsUser: string;
  modifiedBy: string;
  siteTitle: string;
  parentLink: string;
  listID: string;
  listItemID: string;
  spSiteURL: string;
  siteID: string;
  webId: string;
  uniqueID: string;
  spWebUrl: string;
  defaultEncodingURL: string;
  id: string;
}